+++ packages/react-querybuilder/src/types/importExport.ts
@@ -351,7 +351,14 @@ interface ParserCommonOptions {
   fields?: OptionList<FullField> | Record<string, FullField>;
   getValueSources?: (field: string, operator: string) => ValueSources;
   listsAsArrays?: boolean;
+  /**
+   * When true, the generated query will use independent combinators ({@link RuleGroupTypeIC}).
+   */
   independentCombinators?: boolean;
+  /**
+   * When true, a unique `id` will be generated for each rule and group in the query.
+   */
+  generateIDs?: boolean;
 }
 
 /**