+++ packages/react-querybuilder/src/utils/parseMongoDB/parseMongoDB.test.ts
@@ -439,6 +439,16 @@ it('generates query with independent combinators', () => {
   });
 });
 
+it('generates IDs', () => {
+  expect(parseMongoDB({ firstName: 'Steve' }, { generateIDs: true })).toEqual(
+    expect.objectContaining({
+      id: expect.any(String),
+      combinator: 'and',
+      rules: [{ id: expect.any(String), field: 'firstName', operator: '=', value: 'Steve' }],
+    })
+  );
+});
+
 it('handles empty options object', () => {
   expect(parseMongoDB({ f1: 1 }, {})).toEqual({
     combinator: 'and',